requirements.txt
faker
django-stubs[compatible-mypy]
celery
celery-types
celery[redis]
pylint-django
django-extensions
